refactor(sio): split out interpolator and bit utils into their own files
